requirements.txt
requests~=2.32.3
uvicorn~=0.31.0
fastapi~=0.115.0
python-dotenv~=1.0.1